[certs] move attestation key / cert gen code to a new lib
This moves the attestation key / cert gen code to a separate
silicon_creator library so it can be shared across provisioning and
ROM_EXT firmware (in the future).
